feat: добавление запросов на участие, отмена своего запроса, получение списка запросов

+++ ewm-service/src/main/java/ru/practicum/participation_request/dto/EventRequestUpdateDto.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Builder
 public class EventRequestUpdateDto {
 
-    List<Integer> requestIds;
+    List<Long> requestIds;
     RequestStatus status;
 
 }
